requirements.txt
azure-mgmt-cognitiveservices==13.5.0
azure-identity==1.17.0
azure-ai-documentintelligence==1.0.0b3
azure-ai-formrecognizer==3.3.3
azure-cognitiveservices-vision-customvision==3.1.0
openai==1.34.0
streamlit==1.29.0
streamlit_pdf_reader==0.0.5
streamlit_pdf_viewer==0.0.12
pyMuPDF==1.24.5
msg_parser==1.2.0
matplotlib==3.3.2
azure-storage-blob==12.20.0